# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hostreamly Media API",
    "version": "2.0.0",
    "description": "API completa para la plataforma Hostreamly Media con funcionalidades empresariales avanzadas. Esta API proporciona endpoints para gestión de videos, streaming en vivo, analíticas avanzadas, webhooks personalizados, formularios de contacto, enlaces de descarga controlados y mucho más.",
    "termsOfService": "https://hostreamly.com/terms",
    "contact": {
      "name": "Hostreamly Support",
      "email": "support@hostreamly.com",
      "url": "https://hostreamly.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.hostreamly.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error", "message"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Código de error específico",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error",
            "example": "Los datos proporcionados no son válidos"
          },
          "details": {
            "type": "array",
            "description": "Detalles adicionales del error",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "example": "email"
                },
                "message": {
                  "type": "string",
                  "example": "El email debe tener un formato válido"
                }
              }
            }
          }
        },
        "example": {
          "error": "VALIDATION_ERROR",
          "message": "Los datos proporcionados no son válidos",
          "details": [
            {
              "field": "email",
              "message": "El email debe tener un formato válido"
            }
          ]
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/downloads/{token}": {
      "get": {
        "summary": "Obtener información del enlace de descarga",
        "description": "Recupera información detallada sobre un enlace de descarga específico usando su token",
        "tags": ["Downloads"],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "Token único del enlace de descarga",
            "schema": {
              "type": "string",
              "pattern": "^dlt_[a-zA-Z0-9]{16}$",
              "example": "dlt_abcdef1234567890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Información del enlace de descarga",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "videoTitle": {
                          "type": "string",
                          "example": "Mi Video Increíble"
                        },
                        "fileSize": {
                          "type": "integer",
                          "description": "Tamaño del archivo en bytes",
                          "example": 52428800
                        },
                        "duration": {
                          "type": "integer",
                          "description": "Duración del video en segundos",
                          "example": 300
                        },
                        "format": {
                          "type": "string",
                          "example": "mp4"
                        },
                        "quality": {
                          "type": "string",
                          "example": "1080p"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-22T10:30:00Z"
                        },
                        "remainingDownloads": {
                          "type": "integer",
                          "example": 3
                        },
                        "requiresPassword": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Enlace de descarga no encontrado o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Procesar descarga",
        "description": "Inicia la descarga del video asociado al enlace. Requiere validación de contraseña si está configurada.",
        "tags": ["Downloads"],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "description": "Token único del enlace de descarga",
            "schema": {
              "type": "string",
              "pattern": "^dlt_[a-zA-Z0-9]{16}$",
              "example": "dlt_abcdef1234567890"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "Contraseña del enlace (requerida si el enlace está protegido)",
                    "example": "mi-password-seguro"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Descarga iniciada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "downloadUrl": {
                      "type": "string",
                      "format": "uri",
                      "description": "URL directa para descargar el archivo",
                      "example": "https://cdn.hostreamly.com/downloads/temp/vid_1234567890abcdef.mp4?token=temp_xyz789&expires=1642248600"
                    },
                    "expiresIn": {
                      "type": "integer",
                      "description": "Tiempo en segundos hasta que expire la URL de descarga",
                      "example": 3600
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "wrong_password": {
                    "summary": "Contraseña incorrecta",
                    "value": {
                      "error": "INVALID_PASSWORD",
                      "message": "La contraseña proporcionada es incorrecta"
                    }
                  },
                  "ip_not_allowed": {
                    "summary": "IP no permitida",
                    "value": {
                      "error": "IP_NOT_ALLOWED",
                      "message": "Tu dirección IP no está autorizada para descargar este archivo"
                    }
                  },
                  "download_limit_exceeded": {
                    "summary": "Límite de descargas excedido",
                    "value": {
                      "error": "DOWNLOAD_LIMIT_EXCEEDED",
                      "message": "Se ha alcanzado el límite máximo de descargas para este enlace"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Enlace no encontrado o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}